Reuse doubly linked list implementation
